requirements.txt

pandas>=2.0,<3.0
numpy>=1.24,<3.0
scikit-learn>=1.4,<2.0
matplotlib>=3.7,<4.0
joblib>=1.2,<2.0
pyyaml>=6.0,<7.0
tqdm>=4.65,<5.0
gensim>=4.3,<5.0
torch>=2.1,<3.0
transformers>=4.41,<5.0
